separate auth file with user_id handling
